+++ mammoth_wolf_abm/agents/dire_wolf_data.py
@@ -9,7 +9,7 @@ class DireWolfData(AnimalData):
     Attributes:
         ep_gain (int): energy point gained from eating
         max_age (int): maximum age allowed for this agent in years
-        reproductive_age (int): minimum age allowed for reproduction in years
+        reproductive_age (float): minimum age allowed for reproduction in years
         gestation_period (int): gestation period in months
         birth_interval (int): birth_interval in months
         hunt_success_rate (float): Probability of successful hunt in percentage
